+++ packages/client/src/v2-events/hooks/useDefaultValue.ts
@@ -10,22 +10,26 @@
  */
 import { useMemo } from 'react'
 import { useSelector } from 'react-redux'
+import { get } from 'lodash'
 import {
   FieldConfig,
   SystemVariables,
-  InteractiveFieldType,
   SerializedUserField,
   isNonInteractiveFieldType,
   Location,
   FieldType,
   FieldValue,
   EventState,
-  buildFormState
+  buildFormState,
+  FieldGroup,
+  TextField,
+  FieldConfigDefaultValue,
+  isFieldValueWithoutTemplates,
+  isTemplateVariable,
+  mapFieldTypeToZod,
+  compositeFieldTypes
 } from '@opencrvs/commons/client'
-import {
-  getAdminLevelHierarchy,
-  replacePlaceholders
-} from '@client/v2-events/utils'
+import { getAdminLevelHierarchy } from '@client/v2-events/utils'
 import { getOfflineData } from '@client/offline/selectors'
 import { useSystemVariables } from './useSystemVariables'
 import { useLocations } from './useLocations'
@@ -33,6 +37,106 @@ import { useLocations } from './useLocations'
 interface Context extends SystemVariables {
   locations: Location[]
   adminLevelIds: string[]
+}
+
+type FieldsWithDefaultValue =
+  | Extract<FieldConfig, { defaultValue?: unknown }>
+  | FieldGroup
+
+function isFieldWithDefaultValue(
+  field: FieldConfig
+): field is FieldsWithDefaultValue {
+  return 'defaultValue' in field || field.type === FieldType.FIELD_GROUP
+}
+
+function isTextField(field: FieldConfig): field is TextField {
+  return field.type === FieldType.TEXT
+}
+
+/**
+ *
+ * @param fieldType: The type of the field.
+ * @param currentValue: The current value of the field.
+ * @param defaultValue: Configured default value from the country configuration.
+ * @param meta: Metadata fields such as '$user', '$event', and others.
+ *
+ * @returns Resolves template variables in the default value and returns the resolved value.
+ */
+function replacePlaceholders({
+  field,
+  currentValue,
+  defaultValue,
+  systemVariables
+}: {
+  field: FieldsWithDefaultValue
+  currentValue?: FieldValue
+  defaultValue?: FieldConfigDefaultValue
+  systemVariables: SystemVariables
+}): FieldValue | undefined {
+  if (currentValue) {
+    return currentValue
+  }
+
+  if (!defaultValue) {
+    return undefined
+  }
+
+  if (isFieldValueWithoutTemplates(defaultValue)) {
+    return defaultValue
+  }
+
+  if (isTemplateVariable(defaultValue)) {
+    const resolvedValue = get(systemVariables, defaultValue)
+    const validator = mapFieldTypeToZod(field)
+
+    const parsedValue = validator.safeParse(resolvedValue)
+
+    if (parsedValue.success) {
+      return parsedValue.data as FieldValue
+    }
+
+    throw new Error(`Could not resolve ${defaultValue}: ${parsedValue.error}`)
+  }
+
+  if (
+    compositeFieldTypes.some((ft) => ft === field.type) &&
+    typeof defaultValue === 'object'
+  ) {
+    /**
+     * defaultValue is typically an ADDRESS, FILE, or FILE_WITH_OPTIONS.
+     * Some STRING values within the defaultValue object may contain template variables (prefixed with $).
+     */
+    const result = { ...defaultValue }
+
+    // @TODO: This resolves template variables in the first level of the object. In the future, we might need to extend it to arbitrary depth.
+    for (const [key, val] of Object.entries(result)) {
+      if (val && isTemplateVariable(val) && isTextField(field)) {
+        const resolvedValue = get(systemVariables, val)
+        // For now, we only support resolving template variables for text fields.
+        const validator = mapFieldTypeToZod(field)
+        const parsedValue = validator.safeParse(resolvedValue)
+        if (parsedValue.success && parsedValue.data) {
+          result[key] = resolvedValue
+        } else {
+          throw new Error(`Could not resolve ${key}: ${parsedValue.error}`)
+        }
+      }
+    }
+
+    const resultValidator = mapFieldTypeToZod(field)
+    const parsedResult = resultValidator.safeParse(result)
+    if (parsedResult.success) {
+      return result as FieldValue
+    }
+    throw new Error(
+      `Could not resolve ${field.type}: ${JSON.stringify(
+        defaultValue
+      )}. Error: ${parsedResult.error}`
+    )
+  }
+  throw new Error(
+    `Could not resolve ${field.type}: ${JSON.stringify(defaultValue)}`
+  )
 }
 
 function isSerializedUserField(value: unknown): value is SerializedUserField {
@@ -66,7 +170,7 @@ function resolveSerializedUserField(
 }
 
 export function mapFieldToDefaultValue(
-  field: InteractiveFieldType,
+  field: FieldsWithDefaultValue,
   context: Context
 ): FieldValue | undefined {
   if (field.type === FieldType.FIELD_GROUP) {
@@ -146,6 +250,7 @@ export function mapFieldToDefaultValue(
     case FieldType.ID:
     case FieldType.VERIFICATION_STATUS:
     case FieldType.QR_READER:
+    case FieldType.HTTP:
     case FieldType.ID_READER:
     case FieldType.SIGNATURE:
     case FieldType.FILE:
@@ -183,7 +288,7 @@ export function useDefaultValue() {
       return buildFormState(fields, (field) => getDefaultValue(field))
     }
     const field = fieldOrFields
-    if (isNonInteractiveFieldType(field)) {
+    if (!isFieldWithDefaultValue(field)) {
       return
     }
     return mapFieldToDefaultValue(field, {